fix from_qsub
