+++ aiaccel/job/job_status.py
@@ -27,29 +27,3 @@ class JobStatus(IntEnum):
     RUNNING = auto()
     FINISHED = auto()
     ERROR = auto()
-
-
-def from_qsub(status: str) -> JobStatus:
-    """
-    Converts a status string from the qsub command to a JobStatus enum value.
-
-    Args:
-        status (str): The status string from the qsub command.
-
-    Returns:
-        JobStatus: The corresponding JobStatus enum value.
-
-    Raises:
-        ValueError: If the status string is not recognized.
-    """
-    match status:
-        case "r":
-            return JobStatus.RUNNING
-        case "qw" | "h" | "t" | "s" | "S" | "T" | "Rq":
-            return JobStatus.WAITING
-        case "d" | "Rr":
-            return JobStatus.RUNNING
-        case "E":
-            return JobStatus.ERROR
-        case _:
-            raise ValueError(f"Unexpected status: {status}")
